01_injera_quality_inspection/requirements.txt
streamlit
inference-sdk
Pillow